pyproject.toml
[tool.poetry]
name = "cinnamon-rss"
version = "0.1.0"
description = ""
authors = ["Shannon Reid <shannon.victoria.reid@gmail.com>"]
readme = "README.md"
packages = [{include = "cinnamon_rss"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.3"
feedparser = "^6.0.10"
pyfiglet = "^0.8.post1"
python-dotenv = "^1.0.0"
requests = "^2.30.0"
lxml = "^4.9.2"
pymongo = "^4.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
